Ensure we're correctly displaying when vendor files are flashed
With the introduction of the commit below, we added a more specific
Print so that the user can know exactly what is going on, what files
are being flashed onto their device.

https://github.com/DirtyUnicorns/android_build/commit/5d3c766b6c456b522c4e268b24f85aa74f721fe2

The problem here is that with devices that have 'fake treble' or even
a normal vendor partition, the user sees that 'System files' are being
flashed twice. This is misleading and just due to the WriteScript being
called twice, once for the system and once for the vendor.

System
https://github.com/DirtyUnicorns/android_build/blob/p9x/tools/releasetools/ota_from_target_files.py#L889

Vendor
https://github.com/DirtyUnicorns/android_build/blob/p9x/tools/releasetools/ota_from_target_files.py#L901

So what we're doing here is duplicating the WriteScript method and changing
the print so that we're calling on WriteVendorScript for vendor and the regular
WriteScript for system.

Devices that don't have a vendor partition should and won't notice a
different either way.

Change-Id: If488ed574e0f074837b39a6e2e295ff627cd7aa4